requirement.txt
absl-py==0.10.0
addict==2.3.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
anykeystore==0.2
apex==0.1
appdirs==1.4.4
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
av==8.0.3
backcall==0.2.0
black==21.4b2
bleach==1.5.0
braceexpand==0.1.7
cachetools==4.1.1
certifi==2021.5.30
chainer==7.8.1
chainercv==0.13.1
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
clip==1.0
cloudpickle==2.0.0
colorama==0.4.4
cryptacular==1.6.2
cycler==0.10.0
Cython==0.29.28
dataclasses==0.7
datasets==1.16.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.4.0
defusedxml==0.7.1
detectron2==0.6
dill==0.3.4
docker-pycreds==0.4.0
dominate==2.6.0
easydict==1.9
einops==0.3.2
entrypoints==0.4
ExifRead-nocycle==3.0.1
facexlib==0.2.2
filelock==3.4.0
filterpy==1.4.5
fire==0.4.0
flake8==3.8.4
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20220305
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.20
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
huggingface-hub==0.2.1
hupper==1.10.3
hydra-core==1.2.0.dev2
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
img2dataset==1.30.1
importlib-metadata==1.7.0
importlib-resources==5.6.0
imutils==0.5.4
info-nce-pytorch==0.1.4
iopath==0.1.9
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jieba==0.42.1
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.36.0
lmdb==1.2.1
lpips==0.1.3
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.3.2
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.4.8
mmdet==2.23.0
mock==4.0.3
model-index==0.1.11
multidict==5.2.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
natsort==7.1.1
nest-asyncio==1.5.4
networkx==2.5
ninja==1.10.2.3
nltk==3.5
numba==0.53.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233737064/work
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.1.2.30
openmim==0.1.5
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.3
pandas==1.1.5
panopticapi==0.1
parso==0.7.1
PasteDeploy==2.1.1
pathspec==0.9.0
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plaster==1.0
plaster-pastedeploy==0.7
portalocker==2.4.0
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.7.0
pudb==2021.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.6.0
pydensecrf==1.0rc3
pydot==1.4.2
pyflakes==2.2.0
Pygments==2.9.0
pyiqa==0.1.2
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qudida==0.0.4
regex==2020.7.14
repoze.sendmail==4.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.46
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
SQLAlchemy==1.4.27
tabulate==0.8.9
tb-nightly==2.9.0a20220310
tensorboard==1.13.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==1.8
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminaltables==3.1.10
textblob==0.15.3
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.4.12
tokenizers==0.10.3
toml==0.10.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.11.1+cu113
tornado==6.0.4
tqdm==4.62.3
traitlets==4.3.3
transaction==3.0.1
transformers==4.13.0
translationstring==1.4
triton==1.1.1
typed-ast==1.5.4
typing-extensions==3.7.4.3
urllib3==1.25.10
urwid==2.1.2
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
vit-pytorch==0.24.3
wandb==0.12.17
wcwidth==0.2.5
webdataset==0.1.103
WebOb==1.8.7
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==3.0.0
wtforms-recaptcha==0.3.2
xxhash==2.0.2
yacs==0.1.8
yapf==0.32.0
yarl==1.7.2
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
